frontend/feat: [AS] cleanup API index, API dry login, organized files with folders
