+++ frontend/src/components/tables/entries/EntriesTable.jsx
@@ -1,24 +1,27 @@
 import React, { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import Cookies from "universal-cookie";
-import host from "../../assets/strings/host";
-import StudentDropdown from "../../components/dropdowns/StudentDropdown";
+import Search from "../../Search";
+import Statistic from "../../count/Statistic";
+import host from "../../../assets/strings/host";
 
-function SearchEntrySiswa() {
+function EntriesTable() {
   const navigate = useNavigate();
   const cookies = new Cookies();
   const role = cookies.get("role");
   const token = cookies.get("access-token");
+  const max_item = cookies.get("max-item");
 
   const [selectedRows, setSelectedRows] = useState([]);
   const [currentEntry, setCurrentEntry] = useState(0);
   const entryValue = ["permohonan", "pengantaran", "penarikan"];
+  const [pageData, setPageData] = useState();
   const [data, setData] = useState([]);
   const [isDataEdited, setIsDataEdited] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
-  const [selectedValue, setSelectedValue] = useState(null);
-  const value = [{ id: 1, name: "Aan Kurniawan", class: "11 PPLG-1" },{ id: 2, name: "Aaron Ikhwan Saputra", class: "11 PPLG-1" },];
+  const [dataWave, setDataWave] = useState("");
+  const { page } = useParams();
 
   // const fetchWaveData = async () => {
   //   try {
@@ -54,17 +57,21 @@ function SearchEntrySiswa() {
   const fetchDataForEntry = async (entryType) => {
     setLoading(true);
     try {
-      const response = await fetch(`${host}/api/${entryType}?page=0`, {
-        headers: {
-          Authorization: token,
-        },
-      });
+      const response = await fetch(
+        `${host}/api/${entryType}?page=${page}&size=${max_item}`,
+        {
+          headers: {
+            Authorization: token,
+          },
+        }
+      );
       if (!response.ok) {
         throw new Error(`HTTP error: Status ${response.status}`);
       }
       let entryData = await response.json();
       setData(entryData.data.items);
       setIsDataEdited(entryData.data.items.map(() => false));
+      setPageData(entryData.data);
       setError(null);
     } catch (err) {
       console.log("Error fetching data: " + err);
@@ -78,7 +85,7 @@ function SearchEntrySiswa() {
   useEffect(() => {
     console.log(data);
     fetchDataForEntry(entryValue[currentEntry]);
-  }, [currentEntry]);
+  }, [currentEntry, page]);
 
   const deleteEntry = async (id) => {
     try {
@@ -118,6 +125,10 @@ function SearchEntrySiswa() {
 
   function handleSelectTab(index) {
     setCurrentEntry(index);
+  }
+
+  function onAddHandle() {
+    navigate(`/admin/entries/${entryValue[currentEntry]}/add`);
   }
 
   function downloadBlob(blob, name = "file") {
@@ -180,9 +191,14 @@ function SearchEntrySiswa() {
     }
   };
 
+  function handlePageChange(index) {
+    navigate(`/admin/entries/${index}`);
+  }
+
   return (
     <>
-      <StudentDropdown setSelectedValue={setSelectedValue} value={value} />
+      <Search addOnClick={onAddHandle} />
+      <Statistic entryCount={pageData && pageData.total_items} />
       <div className="flex flex-col gap-2">
         <div className="flex justify-between items-center gap-2">
           <div
@@ -202,6 +218,29 @@ function SearchEntrySiswa() {
               </button>
             ))}
           </div>
+          {/* <div className="flex gap-2">
+            <button
+              className={`btn btn-warning btn-sm text-black ${
+                selectedRows.length === 0 ? "opacity-50 cursor-not-allowed" : ""
+              }`}
+              disabled={selectedRows.length === 0}
+            >
+              Export{<span className="hidden lg:block"> yang terpilih</span>}
+            </button>
+            <button
+              className={`btn btn-error btn-sm text-black ${
+                selectedRows.length === 0 ? "opacity-50 cursor-not-allowed" : ""
+              }`}
+              disabled={selectedRows.length === 0}
+              onClick={() =>
+                selectedRows.forEach((rowIndex) =>
+                  deleteEntry(data[rowIndex].id)
+                )
+              }
+            >
+              Delete{<span className="hidden lg:block"> yang terpilih</span>}
+            </button>
+          </div> */}
         </div>
         {loading ? (
           <div>Loading...</div>
@@ -210,7 +249,7 @@ function SearchEntrySiswa() {
             <thead className="bg-neutral">
               <tr className="border-0">
                 <th className="pl-3 pb-2 pr-0 w-0">
-                  <label className="swap">
+                  {/* <label className="swap">
                     <input
                       type="checkbox"
                       onChange={(e) => {
@@ -228,7 +267,8 @@ function SearchEntrySiswa() {
                     <span className="swap-on material-symbols-rounded">
                       check_box
                     </span>
-                  </label>
+                  </label> */}
+                  No
                 </th>
                 <th>Pembimbing</th>
                 <th>Perusahaan</th>
@@ -238,101 +278,105 @@ function SearchEntrySiswa() {
               </tr>
             </thead>
             <tbody className="box-content">
-              {data.map((row, index) => {
-                if (row.company === "Google .Inc") {
-                  return null; // Skip rendering this row
-                }
+              {data.map((row, index) => (
+                <tr key={row.id} className="border-t-2 border-neutral">
+                  <td className="p-3 pb-2">
+                    {/* <label className="swap opacity-60">
+                      <input
+                        type="checkbox"
+                        onChange={() => handleSelectRow(index)}
+                        checked={selectedRows.includes(index)}
+                      />
+                      <span className="swap-off material-symbols-rounded">
+                        check_box_outline_blank
+                      </span>
+                      <span className="swap-on material-symbols-rounded">
+                        check_box
+                      </span>
+                    </label> */}
+                    {index + 1}
+                  </td>
+                  <td>{row.user}</td>
+                  <td>{row.company}</td>
+                  <td>{new Date(row.created_at).toLocaleDateString()}</td>
 
-                return (
-                  <tr key={row.id} className="border-t-2 border-neutral">
-                    <td className="p-3 pb-2">
-                      <label className="swap opacity-60">
-                        <input
-                          type="checkbox"
-                          onChange={() => handleSelectRow(index)}
-                          checked={selectedRows.includes(index)}
-                        />
-                        <span className="swap-off material-symbols-rounded">
-                          check_box_outline_blank
-                        </span>
-                        <span className="swap-on material-symbols-rounded">
-                          check_box
-                        </span>
-                      </label>
-                    </td>
-                    <td>{row.user}</td>
-                    <td>{row.company}</td>
-                    <td>{new Date(row.created_at).toLocaleDateString()}</td>
+                  <td>
+                    {row.verified ? (
+                      <span className="opacity-60">Terverifikasi</span>
+                    ) : (
+                      <span>Belum Terversifikasi</span>
+                    )}
+                  </td>
 
-                    <td>
-                      {row.verified ? (
-                        <span className="opacity-60">Terverifikasi</span>
-                      ) : (
-                        <span>Belum Terversifikasi</span>
-                      )}
-                    </td>
-
-                    <td className="flex flex-row flex-nowrap gap-2">
+                  <td className="flex flex-row flex-nowrap gap-2">
+                    <button
+                      className="btn btn-info btn-xs rounded-lg"
+                      onClick={() => {
+                        navigate(
+                          `/admin/entries/${entryValue[currentEntry]}/${row.id}`
+                        );
+                      }}
+                    >
+                      Detail
+                    </button>
+                    {/* <button
+                      className="btn btn-error btn-xs rounded-lg"
+                      onClick={() => deleteEntry(row.id)}
+                    >
+                      Delete
+                    </button> */}
+                    {row.verified && (
                       <button
-                        className="btn btn-info btn-xs rounded-lg"
-                        onClick={() => {
-                          navigate(
-                            `/admin/entries/${entryValue[currentEntry]}/${row.id}`
-                          );
-                        }}
+                        className="btn btn-warning btn-xs"
+                        onClick={() => onExport(row.id)}
                       >
-                        Detail
+                        Export
                       </button>
-                      <button
-                        className="btn btn-error btn-xs rounded-lg"
-                        onClick={() => deleteEntry(row.id)}
-                      >
-                        Delete
-                      </button>
-                      {row.verified && (
-                        <button
-                          className="btn btn-warning btn-xs"
-                          onClick={() => onExport(row.id)}
-                        >
-                          Export
-                        </button>
-                      )}
-                    </td>
-                  </tr>
-                );
-              })}
+                    )}
+                  </td>
+                </tr>
+              ))}
             </tbody>
           </table>
         )}
-        <div className="flex justify-center items-center gap-2 mt-4">
-          <button className="flex-none btn bg-base-100">
-            <span className="material-symbols-rounded icon-size-20">
-              arrow_back
-            </span>
-          </button>
-          <div className="join flex gap-2">
-            <button className="join-item btn">1</button>
-            <button className="join-item btn">2</button>
-            <button className="join-item btn opacity-50">...</button>
-            <button className="join-item btn">99</button>
-            <button className="join-item btn">100</button>
+        {pageData && (
+          <div className="flex justify-center items-center gap-2 mt-4">
+            <button
+              className="flex-none btn bg-base-100"
+              onClick={() => handlePageChange(pageData.page - 1)}
+              disabled={pageData.page === 0}
+            >
+              <span className="material-symbols-rounded icon-size-20">
+                arrow_back
+              </span>
+            </button>
+            <div className="join flex">
+              {[...Array(pageData.num_pages)].map((_, index) => (
+                <button
+                  key={index}
+                  className={`join-item btn ${
+                    pageData.page === index ? "bg-primary text-base-300" : ""
+                  }`}
+                  onClick={() => handlePageChange(index)}
+                >
+                  {index + 1}
+                </button>
+              ))}
+            </div>
+            <button
+              className="flex-none btn bg-base-100"
+              onClick={() => handlePageChange(pageData.page + 1)}
+              disabled={pageData.page === pageData.num_pages - 1}
+            >
+              <span className="material-symbols-rounded icon-size-20">
+                arrow_forward
+              </span>
+            </button>
           </div>
-          <button className="flex-none btn bg-base-100">
-            <span className="material-symbols-rounded icon-size-20">
-              arrow_forward
-            </span>
-          </button>
-        </div>
-        <button
-          onClick={() => {
-            console.log(data);
-          }}
-        >
-          debug
-        </button>
+        )}
       </div>
     </>
   );
 }
 
-export default SearchEntrySiswa;
+export default EntriesTable;
