+++ frontend/src/components/tables/wave/WaveAddTable.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import host from "../../assets/strings/host";
+import host from "../../../assets/strings/host";
 import Cookies from "universal-cookie";
 import { useNavigate } from "react-router-dom";
