+++ frontend/src/routes/NotFound.jsx
@@ -5,7 +5,7 @@ import NotFoundMessage from "../assets/drawable/NotFoundMessage";
 function NotFound() {
   return (
     <div className="flex h-full flex-col">
-      <Navbar title="Are you lost?" href="/admin/entries" />
+      <Navbar title="Are you lost?" href="/admin/entries/0" />
       <div className="flex-1 flex justify-center items-center bg-base-200"><NotFoundMessage/></div>
     </div>
   );
