--- frontend/src/routes/RefreshToken.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-function RefreshToken() {
-  return (
-    <div></div>
-  )
-}
-
-export default RefreshToken
