--- frontend/src/routes/detail/CompanyAdd.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-import CompanyAddTable from '../../components/tables/CompanyAddTable';
-
-function CompanyAdd() {
-    
-  return (
-    <CompanyAddTable/>
-  )
-}
-
-export default CompanyAdd;
