--- frontend/src/routes/detail/StudentAdd.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-
-
-
-import StudentAddTable from '../../components/tables/StudentAddTable';
-
-function CompanyAdd() {
-    
-  return (
-    <StudentAddTable/>
-  )
-}
-
-export default CompanyAdd;
