--- frontend/src/routes/detail/UserAdd.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-import UserAddTable from '../../components/tables/UserAddTable'
-
-function UserAdd() {
-  return (
-    <>
-      <UserAddTable />
-    </>
-  )
-}
-
-export default UserAdd
