--- frontend/src/routes/detail/WaveAdd.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import WaveAddTable from '../../components/tables/WaveAddTable'
-
-function WaveAdd() {
-  return (
-    <WaveAddTable />
-  )
-}
-
-export default WaveAdd
