+++ frontend/src/routes/entries/EntriesAndDocuments.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import Search from "../components/Search";
-import Statistic from "../components/count/Statistic";
-import EntriesTable from "../components/tables/EntriesTable";
+import Search from "../../components/Search";
+import Statistic from "../../components/count/Statistic";
+import EntriesTable from "../../components/tables/entries/EntriesTable";
 import { useNavigate } from "react-router-dom";
 
 function EntriesAndDocuments() {
