+++ frontend/src/routes/entries/company/Company.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import CompanyTable from "../../components/tables/CompanyTable";
-import Search from "../../components/Search";
+import CompanyTable from "../../../components/tables/company/CompanyTable";
+import Search from "../../../components/Search";
 import { useNavigate } from "react-router-dom";
 
 function Company() {
@@ -10,7 +10,7 @@ function Company() {
   }
   return (
     <>
-      <Search addOnClick={onAddHandle}/>
+      <Search addOnClick={onAddHandle} />
       <CompanyTable />
     </>
   );
