+++ frontend/src/routes/entries/company/CompanyAdd.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import CompanyAddTable from "../../../components/tables/company/CompanyAddTable";
+
+function CompanyAdd() {
+  return <CompanyAddTable />;
+}
+
+export default CompanyAdd;
