+++ frontend/src/routes/entries/student/StudentAdd.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import StudentAddTable from "../../../components/tables/student/StudentAddTable";
+
+function CompanyAdd() {
+  return <StudentAddTable />;
+}
+
+export default CompanyAdd;
