+++ frontend/src/routes/entries/wave/Wave.jsx
@@ -1,7 +1,7 @@
-import React from 'react';
-import Search from '../../components/Search';
-import { useNavigate } from 'react-router-dom';
-import WaveTable from '../../components/tables/WaveTable';
+import React from "react";
+import Search from "../../../components/Search";
+import { useNavigate } from "react-router-dom";
+import WaveTable from "../../../components/tables/wave/WaveTable";
 
 function Wave() {
   const navigate = useNavigate();
