+++ frontend/src/routes/entries/wave/WaveAdd.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import WaveAddTable from "../../../components/tables/wave/WaveAddTable";
+
+function WaveAdd() {
+  return <WaveAddTable />;
+}
+
+export default WaveAdd;
