+++ frontend/src/routes/users/AllUsers.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useTranslation } from "react-i18next";
-import Search from "../components/Search";
-import Users from "../components/tables/UsersTable";
-import StatisticUser from "../components/count/StatisticUser";
+import Search from "../../components/Search";
+import Users from "../../components/tables/users/UsersTable";
+import StatisticUser from "../../components/count/StatisticUser";
 import { useNavigate } from "react-router-dom";
 
 function AllUsers() {
