+++ frontend/src/routes/users/UserAdd.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import UserAddTable from "../../components/tables/users/UserAddTable";
+
+function UserAdd() {
+  return (
+    <>
+      <UserAddTable />
+    </>
+  );
+}
+
+export default UserAdd;
